Numerous changes
Start working on unit-tests for the EleView_display module -- want to be able
to test that all the widgets work, and that they're not getting broken when
changes are made!

Some refactoring of the ElevationDisplay class so it is a little easier to
instantiate for unit-testing.

Fixed a bug that was stopping the QSpinBox from transmitting the change in value
event.